requirements.txt
arrow==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
codecarbon==2.1.0
dataclasses==0.8
docopt-ng==0.7.2
fuzzywuzzy==0.18.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
llvmlite==0.36.0
nltk==3.6.7
numba==0.53.1
numpy==1.19.5
pandas==1.1.5
psutil==5.9.0
py-cpuinfo==8.0.0
pynndescent==0.5.6
pynvml==11.4.1
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.9
zipp==3.6.0
